fix(core): Fix other tests after modifications of error handling
